requirements.txt
Django==5.0.4
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.3.0
reportlab==4.2.2
gunicorn==23.0.0
psycopg2-binary==2.9.9
openrouteservice==2.3.3
